`.zshrc`: Improved load times
Changes:
- Cut down `grep` calls from O(n) to O(1).
- Used array exclusion logic instead of checking on every loop iteration to get necessary arrays for sourcing natively and via omz.
- `npm`: Reduced `diff` usage on load by checking only for the most important line.

+++ custom/plugins/npm/npm.plugin.zsh
@@ -1,7 +1,9 @@
 (( ${+commands[npm]} )) || return 0
 
 # If NPM_CONFIG_USERCONFIG is set, check if file exists and if it contains a given configuration
-if (( ${+NPM_CONFIG_USERCONFIG} )) && ! file_contents_in "$ZDOTDIR/conf/npm/.npmrc" "$NPM_CONFIG_USERCONFIG"; then
+if (( ${+NPM_CONFIG_USERCONFIG} )) && \grep -Eq 'prefix = ${XDG_DATA_HOME}/npm' "$NPM_CONFIG_USERCONFIG" && \
+	! file_contents_in "$ZDOTDIR/conf/npm/.npmrc" "$NPM_CONFIG_USERCONFIG"
+then
 	diff -BNPZbrw --changed-group-format='%>' --unchanged-group-format='' --to-file "$ZDOTDIR/conf/npm/.npmrc" "$NPM_CONFIG_USERCONFIG" > "$XDG_CACHE_HOME"/zsh/npmrc.diff
 	cat "$XDG_CACHE_HOME"/zsh/npmrc.diff >> "$NPM_CONFIG_USERCONFIG"
 	rm -f "$XDG_CACHE_HOME"/zsh/npmrc.diff
